\name{covadap-internal}

\alias{data_preproc}
\alias{p.inverse}
\alias{Imb.m}

\alias{Imb.m.sim}
\alias{print_covadap}
\alias{print_sim}

\alias{print_mixed.covadap}
\alias{print_mixed.sim}

\title{Internal covadap Functions}
\description{
  Internal covadap functions.
}

\value{No return value, called for side effects.}

\note{
  These functions are not to be called by the user (or in some cases are just
  waiting for proper documentation to be written).
}
\keyword{internal}
